+++ iop/fs/devfs/src/devfs.c
@@ -148,7 +148,7 @@ int devfs_fill_dirent(iox_dirent_t *dirent, int devno)
 
    while(dev_scan != NULL)
    {
-     if((curr_no + dev_scan->subdev_count) > devno)
+     if((curr_no + (int)(dev_scan->subdev_count)) > devno)
      {
         for(subdev_loop = 0; subdev_loop < DEVFS_MAX_SUBDEVS; subdev_loop++)
         {
@@ -394,7 +394,7 @@ int devfs_open(iop_file_t *file, const char *name, int mode, int unused)
    {
       char *endp;
       int subdev;
-      int name_len;
+      unsigned int name_len;
 
       name_len = strlen(dev->node.name);
 
@@ -781,7 +781,7 @@ int devfs_lseek(iop_file_t *file, long loc, int whence)
                        break;
         case SEEK_CUR: if(loc > 0)
                          data->loc.loc64 += (u64) loc;
-                       else if(((s64) -loc) < data->loc.loc64)
+                       else if(((s64) -loc) < (s64)(data->loc.loc64))
                          data->loc.loc64 += (s64) loc;
                        break;
         case SEEK_END: if((loc > 0)
@@ -829,7 +829,7 @@ int devfs_lseek64(iop_file_t *file, long long loc, int whence)
                        break;
         case SEEK_CUR: if(loc > 0)
                          data->loc.loc64 += (u64) loc;
-                       else if(((s64) -loc) < data->loc.loc64)
+                       else if(((s64) -loc) < (s64)(data->loc.loc64))
                          data->loc.loc64 += (s64) loc;
                        break;
         case SEEK_END: if((loc > 0)                                                                                  && (dev->subdevs[data->subdev].extent.loc64 >= (u64) loc))
@@ -958,7 +958,7 @@ int devfs_getstat(iop_file_t *file, const char *name, iox_stat_t *stat)
    {
       char *endp;
       int subdev;
-      int name_len;
+      unsigned int name_len;
 
       name_len = strlen(dev->node.name);
